Updated demo 1
> renamed tutorial files to show order they should be presented
> updated dependencies to use specific MSAL & MS Graph packages and latest stable versions
> editted tutorial #1 steps
> updated screenshots
> updated final code demo 1
