--- demos/01-create-app/GraphTutorial/app/src/main/java/com/example/graphtutorial/MainActivity.java
@@ -1,161 +1,0 @@
-package com.example.graphtutorial;
-
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.TextView;
-
-public class MainActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
-    private DrawerLayout mDrawer;
-    private NavigationView mNavigationView;
-    private View mHeaderView;
-    private boolean mIsSignedIn = false;
-    private String mUserName = null;
-    private String mUserEmail = null;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        // Set the toolbar
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        mDrawer = findViewById(R.id.drawer_layout);
-
-        // Add the hamburger menu icon
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, mDrawer, toolbar,
-                R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        mDrawer.addDrawerListener(toggle);
-        toggle.syncState();
-
-        mNavigationView = findViewById(R.id.nav_view);
-
-        // Set user name and email
-        mHeaderView = mNavigationView.getHeaderView(0);
-        setSignedInState(mIsSignedIn);
-
-        // Listen for item select events on menu
-        mNavigationView.setNavigationItemSelectedListener(this);
-
-        // Load the home fragment by default on startup
-        if (savedInstanceState == null) {
-            openHomeFragment(mUserName);
-        }
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
-        // Load the fragment that corresponds to the selected item
-        switch (menuItem.getItemId()) {
-            case R.id.nav_home:
-                openHomeFragment(mUserName);
-                break;
-            case R.id.nav_calendar:
-                openCalendarFragment();
-                break;
-            case R.id.nav_signin:
-                signIn();
-                break;
-            case R.id.nav_signout:
-                signOut();
-                break;
-        }
-
-        mDrawer.closeDrawer(GravityCompat.START);
-
-        return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        if (mDrawer.isDrawerOpen(GravityCompat.START)) {
-            mDrawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    public void showProgressBar()
-    {
-        FrameLayout container = findViewById(R.id.fragment_container);
-        ProgressBar progressBar = findViewById(R.id.progressbar);
-        container.setVisibility(View.GONE);
-        progressBar.setVisibility(View.VISIBLE);
-    }
-
-    public void hideProgressBar()
-    {
-        FrameLayout container = findViewById(R.id.fragment_container);
-        ProgressBar progressBar = findViewById(R.id.progressbar);
-        progressBar.setVisibility(View.GONE);
-        container.setVisibility(View.VISIBLE);
-    }
-
-    // Update the menu and get the user's name and email
-    private void setSignedInState(boolean isSignedIn) {
-        mIsSignedIn = isSignedIn;
-
-        Menu menu = mNavigationView.getMenu();
-
-        // Hide/show the Sign in, Calendar, and Sign Out buttons
-        menu.findItem(R.id.nav_signin).setVisible(!isSignedIn);
-        menu.findItem(R.id.nav_calendar).setVisible(isSignedIn);
-        menu.findItem(R.id.nav_signout).setVisible(isSignedIn);
-
-        // Set the user name and email in the nav drawer
-        TextView userName = mHeaderView.findViewById(R.id.user_name);
-        TextView userEmail = mHeaderView.findViewById(R.id.user_email);
-
-        if (isSignedIn) {
-            // For testing
-            mUserName = "Megan Bowen";
-            mUserEmail = "meganb@contoso.com";
-
-            userName.setText(mUserName);
-            userEmail.setText(mUserEmail);
-        } else {
-            mUserName = null;
-            mUserEmail = null;
-
-            userName.setText("Please sign in");
-            userEmail.setText("");
-        }
-    }
-
-    // Load the "Home" fragment
-    public void openHomeFragment(String userName) {
-        HomeFragment fragment = HomeFragment.createInstance(userName);
-        getSupportFragmentManager().beginTransaction()
-                .replace(R.id.fragment_container, fragment)
-                .commit();
-        mNavigationView.setCheckedItem(R.id.nav_home);
-    }
-
-    // Load the "Calendar" fragment
-    private void openCalendarFragment() {
-        getSupportFragmentManager().beginTransaction()
-                .replace(R.id.fragment_container, new CalendarFragment())
-                .commit();
-        mNavigationView.setCheckedItem(R.id.nav_calendar);
-    }
-
-    private void signIn() {
-        setSignedInState(true);
-        openHomeFragment(mUserName);
-    }
-
-    private void signOut() {
-        setSignedInState(false);
-        openHomeFragment(mUserName);
-    }
-}
