+++ demos/01-create-app/GraphTutorial/app/src/main/java/com/microsoft/graphtutorial/CalendarFragment.java
@@ -1,4 +1,4 @@
-package com.example.graphtutorial;
+package com.microsoft.graphtutorial;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -9,9 +9,10 @@ import android.view.View;
 import android.view.ViewGroup;
 
 public class CalendarFragment extends Fragment {
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_calendar, container, false);
-    }
+  @Nullable
+  @Override
+  public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    return inflater.inflate(R.layout.fragment_calendar, container, false);
+  }
 }
+
