+++ demos/01-create-app/GraphTutorial/app/src/main/java/com/microsoft/graphtutorial/HomeFragment.java
@@ -1,0 +1,52 @@
+package com.microsoft.graphtutorial;
+
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+public class HomeFragment extends Fragment {
+  private static final String USER_NAME = "userName";
+
+  private String mUserName;
+
+  public HomeFragment() {
+
+  }
+
+  public static HomeFragment createInstance(String userName) {
+    HomeFragment fragment = new HomeFragment();
+
+    // Add the provided username to the fragment's arguments
+    Bundle args = new Bundle();
+    args.putString(USER_NAME, userName);
+    fragment.setArguments(args);
+    return fragment;
+  }
+
+  @Override
+  public void onCreate(@Nullable Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    if (getArguments() != null) {
+      mUserName = getArguments().getString(USER_NAME);
+    }
+  }
+
+  @Nullable
+  @Override
+  public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    View homeView = inflater.inflate(R.layout.fragment_home, container, false);
+
+    // If there is a username, replace the "Please sign in" with the username
+    if (mUserName != null) {
+      TextView userName = homeView.findViewById(R.id.home_page_username);
+      userName.setText(mUserName);
+    }
+
+    return homeView;
+  }
+}
