+++ demos/01-create-app/GraphTutorial/app/src/main/res/layout/nav_header.xml
@@ -1,31 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="176dp"
-    android:background="@color/colorPrimary"
-    android:gravity="bottom"
-    android:orientation="vertical"
-    android:padding="16dp"
-    android:theme="@style/ThemeOverlay.AppCompat.Dark">
+  android:layout_width="match_parent"
+  android:layout_height="176dp"
+  android:background="@color/colorPrimary"
+  android:gravity="bottom"
+  android:orientation="vertical"
+  android:padding="16dp"
+  android:theme="@style/ThemeOverlay.AppCompat.Dark">
 
-    <ImageView
-        android:id="@+id/user_profile_pic"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@mipmap/ic_launcher" />
+  <ImageView
+    android:id="@+id/user_profile_pic"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:src="@mipmap/ic_launcher" />
 
-    <TextView
-        android:id="@+id/user_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingTop="8dp"
-        android:text="Test User"
-        android:textAppearance="@style/TextAppearance.AppCompat.Body1" />
+  <TextView
+    android:id="@+id/user_name"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:paddingTop="8dp"
+    android:text="Test User"
+    android:textAppearance="@style/TextAppearance.AppCompat.Body1" />
 
-    <TextView
-        android:id="@+id/user_email"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="test@contoso.com" />
-
+  <TextView
+    android:id="@+id/user_email"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="test@contoso.com" />
 </LinearLayout>
