+++ demos/01-create-app/GraphTutorial/app/src/main/res/menu/drawer_menu.xml
@@ -1,32 +1,31 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
+  xmlns:tools="http://schemas.android.com/tools"
+  tools:showIn="navigation_view">
 
-    <group android:checkableBehavior="single">
-        <item
-            android:id="@+id/nav_home"
-            android:icon="@drawable/ic_menu_home"
-            android:title="Home" />
+  <group android:checkableBehavior="single">
+    <item
+      android:id="@+id/nav_home"
+      android:icon="@drawable/ic_menu_home"
+      android:title="Home" />
 
-        <item
-            android:id="@+id/nav_calendar"
-            android:icon="@drawable/ic_menu_calendar"
-            android:title="Calendar" />
-    </group>
+    <item
+      android:id="@+id/nav_calendar"
+      android:icon="@drawable/ic_menu_calendar"
+      android:title="Calendar" />
+  </group>
 
-    <item android:title="Account">
-        <menu>
-            <item
-                android:id="@+id/nav_signin"
-                android:icon="@drawable/ic_menu_signin"
-                android:title="Sign In" />
-            
-            <item
-                android:id="@+id/nav_signout"
-                android:icon="@drawable/ic_menu_signout"
-                android:title="Sign Out" />
-        </menu>
-    </item>
+  <item android:title="Account">
+    <menu>
+      <item
+        android:id="@+id/nav_signin"
+        android:icon="@drawable/ic_menu_signin"
+        android:title="Sign In" />
 
+      <item
+        android:id="@+id/nav_signout"
+        android:icon="@drawable/ic_menu_signout"
+        android:title="Sign Out" />
+    </menu>
+  </item>
 </menu>
